Verify the downloaded package

+++ src/main/java/com/github/zeropoint/bugfix/Updater.java
@@ -2,6 +2,7 @@ package com.github.zeropoint.bugfix;
 
 import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -9,6 +10,15 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.security.KeyStore;
+import java.security.KeyStoreException;
+import java.security.NoSuchAlgorithmException;
+import java.security.PublicKey;
+import java.security.cert.Certificate;
+import java.security.cert.CertificateException;
+import java.util.jar.JarEntry;
+import java.util.jar.JarFile;
+import java.util.jar.JarInputStream;
 
 import net.minecraft.launchwrapper.Launch;
 
@@ -19,7 +29,7 @@ public class Updater extends Thread {
 
     private static final String ASSET_DOWNLOAD_URL = "https://github.com/wohaopa/ZeroPointServerBugfix/releases/download/%s/ZeroPointBugfix-%s.jar";
     private static final String GITHUB_API_URL = "https://api.github.com/repos/wohaopa/ZeroPointServerBugfix/releases/latest";
-    public static boolean debug = false;
+    public static boolean debug = true;
 
     public static void checkUpdate() {
         new Updater().start();
@@ -54,28 +64,33 @@ public class Updater extends Thread {
 
                 Bugfix.LOG.info("New version available: " + latestVersion);
                 downloadFile(String.format(ASSET_DOWNLOAD_URL, latestVersion, latestVersion));
-                downloaded = true;
-                if (oldMod != null && oldMod.exists()) {
-                    oldMod.deleteOnExit();
-                    Runtime.getRuntime()
-                        .addShutdownHook(new Thread(() -> {
-                            if (oldMod.exists() && !oldMod.delete()) {
-                                File file = new File(SAVE_DIR, "delete.bat");
-                                createDeleteScript(file);
-                                try {
-                                    Runtime.getRuntime()
-                                        .exec(
-                                            "cmd /c start " + file.getAbsolutePath()
-                                                + " \""
-                                                + oldMod.getAbsolutePath()
-                                                + "\"");
-                                } catch (IOException e) {
-                                    e.printStackTrace();
+                if (verify()) {
+                    downloaded = true;
+                    if (oldMod != null && oldMod.exists()) { // It cannot be deleted on windows
+                        oldMod.deleteOnExit();
+                        Runtime.getRuntime()
+                            .addShutdownHook(new Thread(() -> {
+                                if (oldMod.exists() && !oldMod.delete()) {
+                                    File file = new File(SAVE_DIR, "delete.bat");
+                                    createDeleteScript(file);
+                                    try {
+                                        Runtime.getRuntime()
+                                            .exec(
+                                                "cmd /c start " + file.getAbsolutePath()
+                                                    + " \""
+                                                    + oldMod.getAbsolutePath()
+                                                    + "\"");
+                                    } catch (IOException e) {
+                                        e.printStackTrace();
+                                    }
                                 }
-                            }
-                        }));
+                            }));
+                    }
+                    Bugfix.LOG.info("Update downloaded.");
+                } else {
+                    Bugfix.LOG.info("Download file verification failed!");
                 }
-                Bugfix.LOG.info("Update downloaded.");
+
             } else {
                 Bugfix.LOG.info("You are using the latest version.");
             }
@@ -147,6 +162,71 @@ public class Updater extends Thread {
             Bugfix.LOG.info("No file to download. Server replied HTTP code: " + responseCode);
         }
         httpConn.disconnect();
+    }
+
+    private static boolean verify() {
+        String keystoreFile = "ZeroPointBugfixPublic.jks";
+        String keystorePassword = "ZeroPointBugfix";
+        String alias = "ZeroPointBugfix";
+
+        try {
+            JarFile jarFile = new JarFile(oldMod);
+
+            JarEntry jarEntry1 = jarFile.getJarEntry(keystoreFile);
+
+            if (jarEntry1 != null) {
+
+                InputStream inputStream = jarFile.getInputStream(jarEntry1);
+                KeyStore keystore = KeyStore.getInstance(KeyStore.getDefaultType());
+                keystore.load(inputStream, keystorePassword.toCharArray());
+
+                Certificate cert = keystore.getCertificate(alias);
+                if (cert == null) {
+                    Bugfix.LOG.warn("Certificate not found!");
+                    return false;
+                }
+                PublicKey publicKey = cert.getPublicKey();
+
+                JarInputStream jarInputStream = new JarInputStream(new FileInputStream(newMod), true);
+                JarEntry jarEntry2;
+
+                while ((jarEntry2 = jarInputStream.getNextJarEntry()) != null) {
+                    if (jarEntry2.isDirectory()) {
+                        continue;
+                    }
+
+                    byte[] buffer = new byte[8192];
+                    while (jarInputStream.read(buffer, 0, buffer.length) != -1) {
+
+                    }
+
+                    Certificate[] certs = jarEntry2.getCertificates();
+                    if (certs != null) {
+                        for (Certificate jarCert : certs) {
+                            try {
+                                jarCert.verify(publicKey);
+                            } catch (Exception e) {
+                                Bugfix.LOG
+                                    .warn("Validation failed: the signature of {} is invalid.", jarEntry2.getName());
+                                return false;
+                            }
+                        }
+                    } else {
+                        return false;
+                    }
+                }
+
+                jarInputStream.close();
+                inputStream.close();
+                jarFile.close();
+                Bugfix.LOG.info("Successful!");
+                return true;
+            }
+        } catch (IOException | CertificateException | KeyStoreException | NoSuchAlgorithmException e) {
+            return false;
+        }
+
+        return false;
     }
 
     private static final String cmdScript = """
